Refactoring weather tiles request

+++ include/OsmAndCore/TileSqliteDatabase.h
@@ -120,7 +120,7 @@ namespace OsmAnd
 
         bool isEmpty() const;
         bool getTileIds(QList<TileId>& tileIds, ZoomLevel zoom);
-        bool getTilesSize(QList<TileId> tileIds, long long &size, ZoomLevel zoom);
+        bool getTilesSize(QList<TileId> tileIds, uint64_t& size, ZoomLevel zoom);
         bool containsTileData(TileId tileId, ZoomLevel zoom) const;
         bool obtainTileTime(TileId tileId, ZoomLevel zoom, int64_t& outTime) const;
         bool obtainTileData(TileId tileId, ZoomLevel zoom, QByteArray& outData, int64_t* pOutTime = nullptr) const;
