+++ src/TileSqliteDatabase.cpp
@@ -121,7 +121,7 @@ bool OsmAnd::TileSqliteDatabase::getTileIds(QList<TileId>& tileIds, ZoomLevel zo
     return _p->getTileIds(tileIds, zoom);
 }
 
-bool OsmAnd::TileSqliteDatabase::getTilesSize(QList<TileId> tileIds, long long &size, ZoomLevel zoom)
+bool OsmAnd::TileSqliteDatabase::getTilesSize(QList<TileId> tileIds, uint64_t& size, ZoomLevel zoom)
 {
     return _p->getTilesSize(tileIds, size, zoom);
 }
